Main with navbar
