--- src/pages/user.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const User = () => {
-  return <div>user</div>
-}
-
-export default User
